Resolve review comments: Abstract duplicated code in service side sync copying for different blob types.
