+++ lib/TransferControllers/ServerSideSyncCopyControllers/AppendBlobServiceSideSyncCopyController.cs
@@ -21,36 +21,11 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
     /// <summary>
     /// Transfer controller to copy to append blob with AppendBlockFromURL.
     /// </summary>
-    class AppendBlobServiceSideSyncCopyController : TransferControllerBase
+    class AppendBlobServiceSideSyncCopyController : ServiceSideSyncCopyController
     {
-        /// <summary>
-        /// Internal state values.
-        /// </summary>
-        private enum State
-        {
-            FetchSourceAttributes,
-            GetDestination,
-            Copy,
-            Commit,
-            Finished,
-            Error,
-        }
-
-        private AzureBlobLocation sourceLocation = null;
-        private AzureBlobLocation destLocation = null;
-
-        private CloudBlob sourceBlob = null;
-        private CloudAppendBlob destBlob = null;
-
-        private bool hasWork = false;
-
-        private State state;
+        private CloudAppendBlob destAppendBlob = null;
 
         private long blockSize = Constants.DefaultChunkSize;
-
-        private long totalLength;
-
-        private bool gotDestAttributes = false;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="AppendBlobServiceSideSyncCopyController"/> class.
@@ -64,105 +39,12 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
             CancellationToken userCancellationToken)
             : base(scheduler, transferJob, userCancellationToken)
         {
-            if (null == transferJob.Destination)
-            {
-                throw new ArgumentException(
-                    string.Format(
-                        CultureInfo.CurrentCulture,
-                        Resources.ParameterCannotBeNullException,
-                        "Dest"),
-                    "transferJob");
-            }
-
-            this.sourceLocation = this.TransferJob.Source as AzureBlobLocation;
-            this.destLocation = this.TransferJob.Destination as AzureBlobLocation;
-
-            this.sourceBlob = sourceLocation.Blob;
-            this.destBlob = destLocation.Blob as CloudAppendBlob;
-
-            this.state = State.FetchSourceAttributes;
+            this.destBlob = this.destLocation.Blob as CloudAppendBlob;
             this.hasWork = true;
         }
 
-        public override bool HasWork => this.hasWork;
-
-        protected override async Task<bool> DoWorkInternalAsync()
-        {
-            if (!this.TransferJob.Transfer.ShouldTransferChecked)
-            {
-                this.hasWork = false;
-                if (await this.CheckShouldTransfer())
-                {
-                    return true;
-                }
-                else
-                {
-                    this.hasWork = true;
-                    return false;
-                }
-            }
-
-            switch (this.state)
-            {
-                case State.FetchSourceAttributes:
-                    await this.FetchSourceAttributesAsync();
-                    break;
-                case State.GetDestination:
-                    await this.GetDestinationAsync();
-                    break;
-                case State.Copy:
-                    await this.CopyBlockAsync();
-                    break;
-                case State.Commit:
-                    await this.CommitAsync();
-                    break;
-                case State.Finished:
-                case State.Error:
-                default:
-                    break;
-            }
-
-            return (State.Error == this.state || State.Finished == this.state);
-        }
-
-
-        private async Task FetchSourceAttributesAsync()
-        {
-            Debug.Assert(
-                this.state == State.FetchSourceAttributes,
-                "FetchSourceAttributesAsync called, but state isn't FetchSourceAttributes");
-
-            this.hasWork = false;
-            this.StartCallbackHandler();
-
-            AccessCondition accessCondition = null == this.sourceLocation.ETag ?
-                Utils.GenerateConditionWithCustomerCondition(this.sourceLocation.AccessCondition, this.sourceLocation.CheckedAccessCondition) : 
-                Utils.GenerateIfMatchConditionWithCustomerCondition(this.sourceLocation.ETag, this.sourceLocation.AccessCondition, this.destLocation.CheckedAccessCondition);
-
-            try
-            {
-                await this.sourceBlob.FetchAttributesAsync(
-                    accessCondition,
-                    Utils.GenerateBlobRequestOptions(this.sourceLocation.BlobRequestOptions),
-                    Utils.GenerateOperationContext(this.TransferContext));
-            }
-#if EXPECT_INTERNAL_WRAPPEDSTORAGEEXCEPTION
-            catch (Exception ex) when (ex is StorageException || (ex is AggregateException && ex.InnerException is StorageException))
-            {
-                var e = ex as StorageException ?? ex.InnerException as StorageException;
-#else
-            catch (StorageException e)
-            {
-#endif
-                HandleFetchSourceAttributesException(e);
-                throw;
-            }
-
-            this.totalLength = this.sourceBlob.Properties.Length;
-            this.sourceLocation.ETag = this.sourceBlob.Properties.ETag;
-            this.sourceLocation.CheckedAccessCondition = true;
-            // No actual data change has made yet, no need for journal updating.
-            
+        protected override void PostFetchSourceAttributes()
+        {            
             if (0 == this.TransferJob.CheckPoint.EntryTransferOffset)
             {
                 this.state = State.GetDestination;
@@ -182,113 +64,10 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
             this.hasWork = true;
         }
 
-        private static void HandleFetchSourceAttributesException(StorageException e)
-        {
-            // Getting a storage exception is expected if the source doesn't
-            // exist. For those cases that indicate the source doesn't exist
-            // we will set a specific error state.
-            if (e?.RequestInformation?.HttpStatusCode == (int)HttpStatusCode.NotFound)
-            {
-                throw new InvalidOperationException(Resources.SourceDoesNotExistException, e);
-            }
-        }
-
-        private async Task GetDestinationAsync()
+        protected override async Task GetDestinationAsync()
         {
             this.hasWork = false;
-            bool needCreateDestination = true;
-            if (!this.IsForceOverwrite)
-            {
-                if (this.TransferJob.Overwrite.HasValue)
-                {
-                    if (!this.TransferJob.Overwrite.Value)
-                    {
-                        string exceptionMessage = string.Format(CultureInfo.InvariantCulture, Resources.OverwriteCallbackCancelTransferException, this.sourceBlob.Uri.ToString(), this.destBlob.Uri.ToString());
-                        throw new TransferSkippedException(exceptionMessage);
-                    }
-                }
-                else if (!this.destLocation.CheckedAccessCondition && null != this.destLocation.AccessCondition)
-                {
-                    try
-                    {
-                        await this.destBlob.FetchAttributesAsync(
-                            Utils.GenerateConditionWithCustomerCondition(this.destLocation.AccessCondition, false),
-                            Utils.GenerateBlobRequestOptions(this.destLocation.BlobRequestOptions),
-                            Utils.GenerateOperationContext(this.TransferContext),
-                            this.CancellationToken);
-
-                        this.destLocation.CheckedAccessCondition = true;
-
-                        await this.CheckOverwriteAsync(
-                            true,
-                            this.sourceBlob.Uri.ToString(),
-                            this.destBlob.Uri.ToString());
-
-                        this.gotDestAttributes = true;
-                    }
-                    catch (StorageException se)
-                    {
-                        if ((null == se.RequestInformation) || ((int)HttpStatusCode.NotFound != se.RequestInformation.HttpStatusCode))
-                        {
-                            throw;
-                        }
-                    }
-                }
-                else
-                {
-                    AccessCondition accessCondition = new AccessCondition();
-                    accessCondition.IfNoneMatchETag = "*";
-
-                    try
-                    {
-                        await this.destBlob.CreateOrReplaceAsync(
-                            accessCondition,
-                            Utils.GenerateBlobRequestOptions(this.destLocation.BlobRequestOptions),
-                            Utils.GenerateOperationContext(this.TransferContext));
-
-                        needCreateDestination = false;
-                        this.destLocation.CheckedAccessCondition = true;
-                        this.TransferJob.Overwrite = true;
-                        this.TransferJob.Transfer.UpdateJournal();
-                    }
-                    catch (StorageException se)
-                    {
-                        if ((null != se.RequestInformation)
-                            && (((int)HttpStatusCode.PreconditionFailed == se.RequestInformation.HttpStatusCode))
-                                || (((int)HttpStatusCode.Conflict == se.RequestInformation.HttpStatusCode)
-                                    && string.Equals(se.RequestInformation.ErrorCode, "BlobAlreadyExists")))
-                        {
-                            await this.CheckOverwriteAsync(
-                                true,
-                                this.sourceBlob.Uri.ToString(),
-                                this.destBlob.Uri.ToString());
-                        }
-                        else
-                        {
-                            throw;
-                        }
-                    }
-                }
-            }
-            else
-            {
-                this.gotDestAttributes = true;
-            }
-
-            Utils.CheckCancellation(this.CancellationToken);
-
-            if (needCreateDestination)
-            {
-                await this.destBlob.CreateOrReplaceAsync(
-                    Utils.GenerateConditionWithCustomerCondition(this.destLocation.AccessCondition, this.destLocation.CheckedAccessCondition),
-                    Utils.GenerateBlobRequestOptions(this.destLocation.BlobRequestOptions),
-                    Utils.GenerateOperationContext(this.TransferContext));
-
-                this.destLocation.CheckedAccessCondition = true;
-
-                this.TransferJob.Overwrite = true;
-                this.TransferJob.Transfer.UpdateJournal();
-            }
+            await this.CheckAndCreateDestinationAsync();
 
             if (this.TransferJob.CheckPoint.EntryTransferOffset == this.totalLength)
             {
@@ -301,7 +80,7 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
             this.hasWork = true;
         }
 
-        private async Task CopyBlockAsync()
+        protected override async Task CopyChunkAsync()
         {
             long startOffset = this.TransferJob.CheckPoint.EntryTransferOffset;
             this.hasWork = false;
@@ -323,7 +102,7 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
 
             try
             {
-                await this.destBlob.AppendBlockAsync(sourceUri,
+                await this.destAppendBlob.AppendBlockAsync(sourceUri,
                     startOffset,
                     length,
                     null,
@@ -370,6 +149,20 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
             this.hasWork = true;
         }
 
+        protected override async Task CreateDestinationAsync(AccessCondition accessCondition, CancellationToken cancellationToken)
+        {
+            await this.destAppendBlob.CreateOrReplaceAsync(
+                accessCondition,
+                Utils.GenerateBlobRequestOptions(this.destLocation.BlobRequestOptions),
+                Utils.GenerateOperationContext(this.TransferContext),
+                cancellationToken);
+        }
+
+        protected override Task DoPreCopyAsync()
+        {
+            throw new NotImplementedException();
+        }
+
         private async Task<bool> ValidateAppendedChunkAsync(long startOffset, long length)
         {
             await this.destBlob.FetchAttributesAsync(
@@ -401,7 +194,9 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
                     destContentMD5 = eventArgs.RequestInformation.ContentMd5;
                 };
 
-                Task sourceDownloadTask = this.sourceBlob.DownloadRangeToStreamAsync(
+                Task[] downloadTasks = new Task[2];
+
+                downloadTasks[0] = this.sourceBlob.DownloadRangeToStreamAsync(
                     new FakeStream(),
                     startOffset,
                     length,
@@ -412,7 +207,8 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
                     },
                     sourceOperationContext,
                     this.CancellationToken);
-                Task destDownloadTask = this.destBlob.DownloadRangeToStreamAsync(
+
+                downloadTasks[1] = this.destBlob.DownloadRangeToStreamAsync(
                     new FakeStream(),
                     startOffset,
                     length,
@@ -424,9 +220,7 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
                     destOperationContext,
                     this.CancellationToken);
 
-                await sourceDownloadTask;
-                await destDownloadTask;
-
+                await Task.WhenAll(downloadTasks);
                 return (!string.IsNullOrEmpty(sourceConentMD5)) && string.Equals(sourceConentMD5, destContentMD5);
             }
             catch (OperationCanceledException)
@@ -439,55 +233,13 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
             }
         }
 
-        private async Task CommitAsync()
+        protected override async Task CommitAsync()
         {
             Debug.Assert(State.Commit == this.state, "Calling CommitAsync, state should be Commit");
 
             this.hasWork = false;
 
-            BlobRequestOptions blobRequestOptions = Utils.GenerateBlobRequestOptions(this.destLocation.BlobRequestOptions);
-            OperationContext operationContext = Utils.GenerateOperationContext(this.TransferContext);
-
-            if (!this.gotDestAttributes)
-            {
-                await this.destBlob.FetchAttributesAsync(
-                     Utils.GenerateConditionWithCustomerCondition(this.destLocation.AccessCondition),
-                     blobRequestOptions,
-                     operationContext,
-                     this.CancellationToken);
-            }
-
-            var originalMetadata = new Dictionary<string, string>(this.destBlob.Metadata);
-            var sourceProperties = this.sourceBlob.Properties;
-
-            Utils.SetAttributes(this.destBlob,
-                 new Attributes()
-                 {
-                     CacheControl = sourceProperties.CacheControl,
-                     ContentDisposition = sourceProperties.ContentDisposition,
-                     ContentEncoding = sourceProperties.ContentEncoding,
-                     ContentLanguage = sourceProperties.ContentLanguage,
-                     ContentMD5 = sourceProperties.ContentMD5,
-                     ContentType = sourceProperties.ContentType,
-                     Metadata = this.sourceBlob.Metadata,
-                     OverWriteAll = true
-                 });
-            await this.SetCustomAttributesAsync(this.destBlob);
-
-            await this.destBlob.SetPropertiesAsync(
-                Utils.GenerateConditionWithCustomerCondition(this.destLocation.AccessCondition),
-                blobRequestOptions,
-                operationContext,
-                this.CancellationToken);
-
-            if (!originalMetadata.DictionaryEquals(this.destBlob.Metadata))
-            {
-                await this.destBlob.SetMetadataAsync(
-                    Utils.GenerateConditionWithCustomerCondition(this.destLocation.AccessCondition),
-                    blobRequestOptions,
-                    operationContext,
-                    this.CancellationToken);
-            }
+            await this.CommonCommitAsync();
 
             this.SetFinish();
         }
