+++ lib/TransferControllers/TransferControllerBase.cs
@@ -128,48 +128,6 @@ namespace Microsoft.Azure.Storage.DataMovement.TransferControllers
             set;
         }
 
-        public static TransferControllerBase GenerateTransferConstroller(
-            TransferScheduler transferScheduler,
-            TransferJob transferJob,
-            CancellationToken cancellationToken)
-        {
-            TransferControllerBase controller = null;
-            switch (transferJob.Transfer.TransferMethod)
-            {
-                case TransferMethod.SyncCopy:
-                        controller = new SyncTransferController(transferScheduler, transferJob, cancellationToken);
-                    break;
-
-                case TransferMethod.ServiceSideAsyncCopy:
-                    controller = AsyncCopyController.CreateAsyncCopyController(transferScheduler, transferJob, cancellationToken);
-                    break;
-
-                case TransferMethod.ServiceSideSyncCopy:
-                    if (transferJob.Destination.Instance is CloudBlockBlob)
-                    {
-                        controller = new BlockBlobServiceSideSyncCopyController(transferScheduler, transferJob, cancellationToken);
-                    }
-                    else if (transferJob.Destination.Instance is CloudAppendBlob)
-                    {
-                        controller = new AppendBlobServiceSideSyncCopyController(transferScheduler, transferJob, cancellationToken);
-                    }
-                    else if (transferJob.Destination.Instance is CloudPageBlob)
-                    {
-                        controller = new PageBlobServiceSideSyncCopyController(transferScheduler, transferJob, cancellationToken);
-                    }
-                    else
-                    {
-                        //TODO throw out exception
-                    }
-                    break;
-                case TransferMethod.DummyCopy:
-                    controller = new DummyTransferController(transferScheduler, transferJob, cancellationToken);
-                    break;
-            }
-
-            return controller;
-        }
-
         /// <summary>
         /// Gets scheduler object which creates this object.
         /// </summary>
